// examples/mock-http-server/src/types/generated.ts
export type DeliveryDiscountCandidate = {
  /** The discount code associated with this discount candidate, for code-based discounts. */
  associatedDiscountCode?: InputMaybe<AssociatedDiscountCode>;
  /** The discount message. */
  message?: InputMaybe<Scalars["String"]["input"]>;
  /** The targets of the delivery discount candidate. */
  targets: Array<DeliveryDiscountCandidateTarget>;
  /** The value of the delivery discount candidate. */
  value: DeliveryDiscountCandidateValue;
};
// ...
export enum DeliveryDiscountSelectionStrategy {
  /**
   * Apply all delivery discount candidates with conditions that are satisfied. This does not override
   * discount combination or stacking rules.
   */
  All = "ALL",
}

/** A group of delivery discounts that share a selection strategy. */
export type DeliveryDiscounts = {
  /** The list of delivery discount candidates to be applied. */
  candidates: Array<DeliveryDiscountCandidate>;
  /** The strategy that's applied to the list of discounts. */
  selectionStrategy: DeliveryDiscountSelectionStrategy;
};
// ...
export type DeliveryOperation =
  /** A group of delivery discounts that share a selection strategy. */
  | {
      deliveryDiscountsAdd: DeliveryDiscounts;
      enteredDiscountCodesAccept?: never;
    } /**
   * A list of valid discount codes that correspond to external discounts. This can
   * only be used by Functions with network access.
   */
  | {
      deliveryDiscountsAdd?: never;
      enteredDiscountCodesAccept: ValidDiscountCodes;
    };
// ...
export type ValidDiscountCode = {
  /** The discount code. */
  code: Scalars["String"]["input"];
};

/**
 * A list of valid discount codes that correspond to external discounts. This can
 * only be used by Functions with network access.
 */
export type ValidDiscountCodes = {
  /** The list of valid discount codes. */
  codes: Array<ValidDiscountCode>;
};